fix: improve cursor position handling for sixels
By sending DECSC/DECRC before/after drawing the sixel data, we restore
the cursor to the position of the end of the placeholder. This improves
the situation with missing frames and padding a lot, even though some
issues remain when images higher than the screen are rendered next to
each other.