fix: prevent image from not being rendered
If Textual doesn't know the container height when initially rendering an
image, it passes 0 as height to get_content_width/get_content_height.
If using `auto` values for width/height in the image this may lead to
the image not being rendered until the first redraw. We're passing some
huge numbers as container height to the image size calculation now in
this case, to prevent this.